Implementacion del envio de correo electronico

+++ app/Http/Controllers/ComentarioController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Comentario;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ComentarioController extends Controller
 {
@@ -24,7 +25,7 @@ class ComentarioController extends Controller
         //     ->where('nombre', 'like', '%2')
         //     ->get();
 
-        $comentarios = Auth::user()->comentarios;
+        $comentarios = Comentario::all();
 
         return view('comentarios/comentarioIndex', compact('comentarios'));
     }
@@ -70,6 +71,8 @@ class ComentarioController extends Controller
      */
     public function edit(Comentario $comentario)
     {
+        $this->authorize('update', $comentario);
+
         return view('comentarios.comentarioEdit', compact('comentario'));
     }
 
@@ -78,6 +81,8 @@ class ComentarioController extends Controller
      */
     public function update(Request $request, Comentario $comentario)
     {
+        $this->authorize('update', $comentario);
+
         $request->validate([
             'nombre' => 'required|max:255',
             'correo' => ['required', 'email', 'max:255'],
@@ -101,6 +106,8 @@ class ComentarioController extends Controller
      */
     public function destroy(Comentario $comentario)
     {
+        $this->authorize('delete', $comentario);
+
         $comentario->delete();
         return redirect()->route('comentario.index');
     }
